Adjustments of .c files. Adjustment of load_map and choose_menu_item and its test files.
