+++ src/choose_menu_item.c
@@ -1,7 +1,6 @@
 /**
  * @file choose_menu_item.c
- * @author Pari
- * @brief Chooses menu items and test it.
+ * \brief Chooses menu items and test it.
  *
  * This file chooses a menu item and test if the function is working.
  */
@@ -10,11 +9,25 @@
 #include "menu.h"
 
 
+void user_manual() {
+	/*TODO*/
+}
+
+void single_player() {
+	/*TODO*/
+}
+
+void credit() {
+	/*TODO*/
+}
+
+void exit_game() {
+	/*TODO*/
+}
 /**
- * @brief chooses a menu item.
+ * \brief chooses a menu item.
  *
  * This function choose one of the menu items based on the selector.
- *
  * @param [in] menu which is a structure.
  * @return void.
  */
@@ -39,20 +52,4 @@ void choose_menu_item(menu_t menu) {
 	}
 }
 
-/**
- * @brief tests choose_menu_item() functions.
- *
- * This is a test function for choose_menu_item() and check if it performs correctly.
- * It has no input.
- *maa
- * @return 0 in success
- */
-int test_choose_menu_item() {
-	menu_t menu;
-	for (int item = 0; item < MENU_ITEM_NUM_OF_ITEMS; item++)
-	{
-		menu.selector = item;
-		choose_menu_item(menu);
-	}
-	return 0;
-}
+
