+++ test/load_map_test.c
@@ -1,10 +1,6 @@
 /**
- * @file load_map.c
- * @author Pari
- * @brief Loads the map.
- *
- * This file implements map loading functionality
- * and a basic testcase to check if it works correctly.
+ * @file load_map_test.c
+ * \brief A testcase to check if load_map() works correctly.
  */
 #include <stdio.h>
 #include <string.h>
@@ -12,67 +8,7 @@
 #include "map.h"
 
 /**
- * @brief loads the map.
- *
- * This function parses the map file and initializes the map structure using input file.
- *
- * @param [in] file_path input file path.
- * @return filled map structure.
- */
-map_t load_map(char* file_path) {
-
-	FILE* map_file = fopen((const char *)file_path, "r");
-
-	if (map_file == NULL) {
-		printf("Error opening file.\n");
-		assert(0);
-	}
-	map_t map = {0};
-
-	fscanf(map_file, "space = %d %d %d %d\n",
-		   &map.space.x_min,
-		   &map.space.x_max,
-		   &map.space.y_min,
-		   &map.space.y_max);
-
-	fscanf(map_file, "goal = %d %d\n",
-		   &map.goal.pos.x,
-		   &map.goal.pos.y);
-
-	fscanf(map_file, "player = %d %d %d\n",
-		   &map.player.current_pos.x,
-		   &map.player.current_pos.y,
-		   &map.player.heart);
-
-	fscanf(map_file,"number_of_barriers = %d\n", &map.number_of_barriers);
-	assert(map.number_of_barriers <= MAP_MAX_NUM_OF_BARRIERS);
-
-	for (int i = 0; i < map.number_of_barriers; i++) {
-		fscanf(map_file, "barrier = %d %d %d %d\n",
-			   &map.barrier[i].current_pos.x,
-			   &map.barrier[i].current_pos.y,
-			   &map.barrier[i].length,
-			   &map.barrier[i].current_dir);
- 	}
-
-
-	fscanf(map_file, "number_of_arrows = %d\n", &map.number_of_arrows);
-	assert(map.number_of_arrows <= MAP_MAX_NUM_OF_ARROWS);
-
-	for (int i = 0; i < map.number_of_arrows; i++) {
-		fscanf(map_file, "arrow = %d %d %d %d\n",
-			   &map.arrow[i].current_pos.x,
-			   &map.arrow[i].current_pos.y,
-			   &map.arrow[i].speed,
-			   &map.arrow[i].direction);
-	}
-
-	fclose(map_file);
-	return map;
-}
-
-/**
- * @brief test load_map() function.
+ * \brief load_map_test() function.
  *
  * This is the test function for load_map(). It has no input or output variable.
  * This function creates a map_file and defines the map requirements into that.
@@ -80,7 +16,7 @@ map_t load_map(char* file_path) {
  *
  * @return 0 in success.
  */
-int test_load_map(){
+int load_map_test(){
 
 	char file_path[1000] = "./test_file.txt";
 	FILE* map_file = fopen((const char*)file_path, "w");
@@ -125,3 +61,5 @@ int test_load_map(){
 
 	return 0;
 }
+
+
